Remove package-lock.json, which causes problems when yarn is used
In our setup, we need to use npm first to install yarn; newer versions
of node (we upgraded to 10 to migrate away from bower), create a file
package-lock.json, which stores hashes of packages installed by npm.
When yarn then install the actual deps, it installs them into
node_modules. If npm runs again, it checks for package-lock.json
presence, it detects that the hash reported there for yarn, the only
package it installs doesn't match with the status of the node_modules
dir, which has many more packages, and removes all other packages
apart from yarn; yarn then checks that the packages in package.json
match with those in its own lock file, yarn.lock, and it doesn't
re-install the packages that npm removed.

The 'solution' is to delete package-lock.json after every npm
execution, or to always use --no-package-lock.